verdict fixes from 7d5ce210 (884/1366): c910 loop cond (u32)i<0x80 (jae), eb50 drop named offset local (orig: this-spill@ebp-4 first, implicit fild temp@ebp-8), FUN_00464070 revert variadic -> plain member (DrawNoRotation proves thiscall; variadic fell back to cdecl-this-on-stack; watch rotate to 4 Player near-misses)

+++ src/Player.cpp
@@ -400,7 +400,7 @@ void Player::FUN_0040c910()
         }
     }
 
-    for (entry = bomb + 0x16f4c, i = 0x10; i < 0x80; i++, entry += 0x16f0)
+    for (entry = bomb + 0x16f4c, i = 0x10; (u32)i < 0x80; i++, entry += 0x16f0)
     {
         if (*(i32 *)entry == 0)
         {
